Rename executable assembly to KurrentDB
