+++ src/KurrentDB/KurrentDB.csproj
