+++ src/KurrentDB/logconfig.json
